REFACT : model moved to domain - gradle warnings handle - bug: Row too big to fit into CursorWindow requiredPos=0, totalRows=1
