+++ app/build.gradle
@@ -10,8 +10,8 @@ android {
 
     defaultConfig {
         applicationId "com.alphaomardiallo.parisforkids"
-        minSdk 21
-        targetSdk 32
+        minSdk 24
+        targetSdk 33
         versionCode 1
         versionName "1.0"
 
@@ -75,35 +75,33 @@ dependencies {
     implementation "androidx.compose.ui:ui:$compose_version"
     implementation "androidx.compose.material:material:$compose_version"
     implementation "androidx.compose.ui:ui-tooling-preview:$compose_version"
-    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.3.1'
-    implementation 'androidx.activity:activity-compose:1.3.1'
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    implementation "androidx.lifecycle:lifecycle-runtime-ktx:$lifecycleRuntime_version"
+    implementation "androidx.activity:activity-compose:$composeActivity_version"
+
+    testImplementation "junit:junit:$junit_version"
+    androidTestImplementation "androidx.test.ext:junit:$junitTest_version"
+    androidTestImplementation "androidx.test.espresso:espresso-core:$espressoCore_version"
     androidTestImplementation "androidx.compose.ui:ui-test-junit4:$compose_version"
     debugImplementation "androidx.compose.ui:ui-tooling:$compose_version"
     debugImplementation "androidx.compose.ui:ui-test-manifest:$compose_version"
 
     //TEST
-    testImplementation 'org.robolectric:robolectric:4.8'
-    testImplementation "androidx.test.ext:truth:1.4.0"
-    testImplementation "org.mockito:mockito-core:4.3.1"
-    androidTestImplementation "androidx.test.ext:truth:1.4.0"
+    testImplementation "org.robolectric:robolectric:$robolectric_version"
+    testImplementation "androidx.test.ext:truth:$truth_version"
+    testImplementation "org.mockito:mockito-core:$mockito_version"
+    androidTestImplementation "androidx.test.ext:truth:$truth_version"
 
     //DAGGER - HILT
-    implementation 'com.google.dagger:hilt-android:2.42'
-    kapt 'com.google.dagger:hilt-compiler:2.42'
+    implementation "com.google.dagger:hilt-android:$hilt_version"
+    kapt "com.google.dagger:hilt-compiler:$hilt_version"
     // For instrumentation tests
-    androidTestImplementation 'com.google.dagger:hilt-android-testing:2.42'
-    kaptAndroidTest 'com.google.dagger:hilt-compiler:2.42'
+    androidTestImplementation "com.google.dagger:hilt-android-testing:$hilt_version"
+    kaptAndroidTest "com.google.dagger:hilt-compiler:$hilt_version"
     // For local unit tests
-    testImplementation 'com.google.dagger:hilt-android-testing:2.42'
-    kaptTest 'com.google.dagger:hilt-compiler:2.42'
+    testImplementation "com.google.dagger:hilt-android-testing:$hilt_version"
+    kaptTest "com.google.dagger:hilt-compiler:$hilt_version"
 
     //LIFECYCLE
-    def lifecycle_version = "2.6.0-alpha01"
-    def arch_version = "2.1.0"
-
     // ViewModel
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version"
     // ViewModel utilities for Compose
@@ -119,7 +117,6 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-common-java8:$lifecycle_version"
 
     //NAVIGATION
-    def nav_version = "2.5.1"
     implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
     implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
     // Feature module Support
@@ -130,47 +127,46 @@ dependencies {
     implementation "androidx.navigation:navigation-compose:$nav_version"
 
     //PREFERENCE DATASTORE
-    implementation "androidx.datastore:datastore-preferences:1.0.0"
+    implementation "androidx.datastore:datastore-preferences:$datastore_version"
 
     //WORK MANAGER
     def work_version = "2.7.1"
     implementation "androidx.work:work-runtime-ktx:$work_version"
 
     //COROUTINES
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.1'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.1'
-    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.1"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutines_version"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutines_version"
+    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:$coroutines_version"
 
     //RETROFIT - API CLIENT
-    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
-    implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
-    implementation 'com.google.code.gson:gson:2.9.0'
+    implementation "com.squareup.retrofit2:retrofit:$retrofit_version"
+    implementation "com.squareup.retrofit2:converter-gson:$retrofit_version"
+    implementation "com.google.code.gson:gson:$gson_version"
 
     //OKHTTP
-    implementation 'com.squareup.okhttp3:okhttp:5.0.0-alpha.2'
-    implementation "com.squareup.okhttp3:logging-interceptor:5.0.0-alpha.2"
-    testImplementation 'com.squareup.okhttp3:mockwebserver:4.10.0'
-    androidTestImplementation "com.squareup.okhttp3:mockwebserver:4.10.0"
+    implementation "com.squareup.okhttp3:okhttp:$okhttp_version"
+    implementation "com.squareup.okhttp3:logging-interceptor:$okhttp_version"
+    testImplementation "com.squareup.okhttp3:mockwebserver:$mockwebserver_version"
+    androidTestImplementation "com.squareup.okhttp3:mockwebserver:$mockwebserver_version"
 
     //EASY PERMISSIONS - runtime permissions library
-    implementation 'com.vmadalin:easypermissions-ktx:1.0.0'
+    implementation "com.vmadalin:easypermissions-ktx:$easypermission_version"
 
     //FIREBASE
     //Firebase BOM
-    implementation platform('com.google.firebase:firebase-bom:30.2.0')
+    implementation platform("com.google.firebase:firebase-bom:firebaseBom_version")
     //Firebase Analytics
     implementation 'com.google.firebase:firebase-analytics-ktx'
     //Firestore database
     implementation 'com.google.firebase:firebase-firestore-ktx'
     //Firebase Auth
     implementation 'com.google.firebase:firebase-auth-ktx'
-    implementation 'com.firebaseui:firebase-ui-auth:8.0.1'
+    implementation "com.firebaseui:firebase-ui-auth:$firebaseUiAuth"
     //Firebase cloud storage
     implementation 'com.google.firebase:firebase-storage'
 
-    def room_version = "2.4.3"
-
-    implementation"androidx.room:room-ktx:$room_version"
+    //ROOM DATABASE
+    implementation "androidx.room:room-ktx:$room_version"
     implementation "androidx.room:room-runtime:$room_version"
     annotationProcessor "androidx.room:room-compiler:$room_version"
     // To use Kotlin annotation processing tool (kapt)
@@ -179,14 +175,13 @@ dependencies {
     testImplementation "androidx.room:room-testing:$room_version"
 
     //COIL
-    implementation "io.coil-kt:coil:2.1.0"
-    implementation "io.coil-kt:coil-compose:2.1.0"
+    implementation "io.coil-kt:coil:$coil_version"
+    implementation "io.coil-kt:coil-compose:$coil_version"
 
     //ICONS
     implementation "androidx.compose.material:material-icons-extended:$compose_version"
 
     //PAGING 3
-    def paging_version = "3.1.1"
     implementation "androidx.paging:paging-runtime:$paging_version"
-    implementation "androidx.paging:paging-compose:1.0.0-alpha16"
+    implementation "androidx.paging:paging-compose:$pagingCompose_version"
 }
