+++ app/src/main/java/com/alphaomardiallo/parisforkids/data/remote/RetrofitQueFaireAParis.kt
@@ -1,6 +1,6 @@
 package com.alphaomardiallo.parisforkids.data.remote
 
-import com.alphaomardiallo.parisforkids.data.model.queFaireAParis.ResponseQueFaireAParis
+import com.alphaomardiallo.parisforkids.domain.model.queFaireAParis.ResponseQueFaireAParis
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
