+++ app/src/main/java/com/alphaomardiallo/parisforkids/data/repository/events/EventsRepository.kt
@@ -1,6 +1,6 @@
 package com.alphaomardiallo.parisforkids.data.repository.events
 
-import com.alphaomardiallo.parisforkids.data.model.queFaireAParis.Events
+import com.alphaomardiallo.parisforkids.domain.model.queFaireAParis.Events
 import kotlinx.coroutines.flow.Flow
 
 interface EventsRepository {
