+++ app/src/main/java/com/alphaomardiallo/parisforkids/data/repository/parisWeather/ParisWeatherRepository.kt
@@ -1,6 +1,6 @@
 package com.alphaomardiallo.parisforkids.data.repository.parisWeather
 
-import com.alphaomardiallo.parisforkids.data.model.weather.ResponseWeather
+import com.alphaomardiallo.parisforkids.domain.model.weather.ResponseWeather
 import retrofit2.Response
 
 interface ParisWeatherRepository {
