+++ app/src/main/java/com/alphaomardiallo/parisforkids/data/repository/queFaireAParis/QueFaireAParisRepository.kt
@@ -1,6 +1,6 @@
 package com.alphaomardiallo.parisforkids.data.repository.queFaireAParis
 
-import com.alphaomardiallo.parisforkids.data.model.queFaireAParis.ResponseQueFaireAParis
+import com.alphaomardiallo.parisforkids.domain.model.queFaireAParis.ResponseQueFaireAParis
 import retrofit2.Response
 
 interface QueFaireAParisRepository {
