+++ app/src/main/java/com/alphaomardiallo/parisforkids/data/repository/queFaireAParis/QueFaireAParisRepositoryImp.kt
@@ -1,6 +1,6 @@
 package com.alphaomardiallo.parisforkids.data.repository.queFaireAParis
 
-import com.alphaomardiallo.parisforkids.data.model.queFaireAParis.ResponseQueFaireAParis
+import com.alphaomardiallo.parisforkids.domain.model.queFaireAParis.ResponseQueFaireAParis
 import com.alphaomardiallo.parisforkids.data.remote.RetrofitQueFaireAParis
 import retrofit2.Response
 import javax.inject.Inject
