+++ app/src/main/java/com/alphaomardiallo/parisforkids/data/repository/weather/WeatherRepository.kt
@@ -1,6 +1,6 @@
 package com.alphaomardiallo.parisforkids.data.repository.weather
 
-import com.alphaomardiallo.parisforkids.data.model.weather.Weather
+import com.alphaomardiallo.parisforkids.domain.model.weather.Weather
 import kotlinx.coroutines.flow.Flow
 
 interface WeatherRepository {
