+++ app/src/main/java/com/alphaomardiallo/parisforkids/data/repository/weather/WeatherRepositoryImp.kt
@@ -1,7 +1,7 @@
 package com.alphaomardiallo.parisforkids.data.repository.weather
 
 import com.alphaomardiallo.parisforkids.data.local.WeatherDAO
-import com.alphaomardiallo.parisforkids.data.model.weather.Weather
+import com.alphaomardiallo.parisforkids.domain.model.weather.Weather
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
