+++ app/src/main/java/com/alphaomardiallo/parisforkids/domain/model/sanisette/Fields.kt
@@ -1,4 +1,4 @@
-package com.alphaomardiallo.parisforkids.data.model.sanisette
+package com.alphaomardiallo.parisforkids.domain.model.sanisette
 
 data class Fields(
     val acces_pmr: String,
