+++ app/src/main/java/com/alphaomardiallo/parisforkids/domain/model/sanisette/Geometry.kt
@@ -1,4 +1,4 @@
-package com.alphaomardiallo.parisforkids.data.model.sanisette
+package com.alphaomardiallo.parisforkids.domain.model.sanisette
 
 data class Geometry(
     val coordinates: List<Double>,
