+++ app/src/main/java/com/alphaomardiallo/parisforkids/domain/model/weather/Weather.kt
@@ -1,4 +1,4 @@
-package com.alphaomardiallo.parisforkids.data.model.weather
+package com.alphaomardiallo.parisforkids.domain.model.weather
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
